exynos4: hotplug: Gingerbread legacy hotplug
It seems like the existing hotplug algorithms in the ICS kernel
are not quite mature enough for production yet.  See all the hackery
needed to not get stand-hotplug to turn off the boot CPU...

This forward-ports a tuned version of the Gingerbread legacy
hotplug algorithm (includes support for screen-off profiles and
minimum hotplug frequency).

Change-Id: I22a999202e752308446244c272b23960a7b108ee
Signed-off-by: Andrew Dodd <atd7@cornell.edu>

Conflicts:

	arch/arm/mach-exynos/Kconfig